Reorganizing the project
prioritizing inline structure and avoiding
excessive use of packages +
deleting unnecessary config files
