+++ tmail-backend/tmail-third-party/ai-bot/src/main/scala/com/linagora/tmail/jmap/aibot/AiBotCapability.scala
@@ -15,22 +15,16 @@
  * PURPOSE. See the GNU Affero General Public License for          *
  * more details.                                                   *
  * ****************************************************************** */
-package com.linagora.tmail.jmap.core
+package com.linagora.tmail.jmap.aibot
 
-import com.linagora.tmail.jmap.core.CapabilityIdentifier.LINAGORA_AIBOT
 import com.google.inject.AbstractModule
 import com.google.inject.multibindings.{Multibinder, ProvidesIntoSet}
-import com.linagora.tmail.jmap.method.AiBotSuggestionMethod
+import com.linagora.tmail.jmap.aibot.CapabilityIdentifier.LINAGORA_AIBOT
 import eu.timepit.refined.auto._
 import org.apache.james.jmap.core.CapabilityIdentifier.CapabilityIdentifier
-import org.apache.james.jmap.core.CapabilityIdentifier.JMAP_CORE
 import org.apache.james.jmap.core.{Capability, CapabilityFactory, CapabilityProperties, UrlPrefixes}
-import org.apache.james.jmap.core.Invocation.MethodName
-import org.apache.james.jmap.method.{InvocationWithContext, Method}
-import org.apache.james.mailbox.MailboxSession
-import org.reactivestreams.Publisher
+import org.apache.james.jmap.method.Method
 import play.api.libs.json.{JsObject, Json}
-import reactor.core.scala.publisher.SMono
 
 object CapabilityIdentifier {
   val LINAGORA_AIBOT: CapabilityIdentifier = "com:linagora:params:jmap:aibot"
@@ -64,4 +58,3 @@ class AiBotMethodModule extends AbstractModule {
       .to(classOf[AiBotSuggestionMethod])
   }
 }
-
