+++ tmail-backend/tmail-third-party/ai-bot/src/main/scala/com/linagora/tmail/jmap/aibot/AiBotSuggestReplyRequest.scala
@@ -15,24 +15,26 @@
  * PURPOSE. See the GNU Affero General Public License for          *
  * more details.                                                   *
  * ****************************************************************** */
-package com.linagora.tmail.jmap.json
+package com.linagora.tmail.jmap.aibot
 
-import com.linagora.tmail.jmap.mail.{AiBotSuggestReplyRequest, AiBotSuggestReplyResponse}
-import play.api.libs.json._
+import org.apache.james.jmap.core.AccountId
+import org.apache.james.jmap.core.Id.Id
+import org.apache.james.jmap.method.WithAccountId
 
+case class AiBotSuggestReplyRequest(
+                                     accountId: AccountId,
+                                     emailId: Option[Id],
+                                     userInput: String
+                                   )extends WithAccountId
 
-object AiBotSerializer {
-
-  private implicit val identityGetRequestReads: Reads[AiBotSuggestReplyRequest] = Json.reads[AiBotSuggestReplyRequest]
-  private implicit val AiBotSuggestReplyResponseWrites:  OWrites[AiBotSuggestReplyResponse] = Json.writes[AiBotSuggestReplyResponse]
-  private implicit val identityGetResponseReads: Reads[AiBotSuggestReplyResponse] = Json.reads[AiBotSuggestReplyResponse]
-
-  def deserializeRequest(json: JsValue): JsResult[AiBotSuggestReplyRequest] = {
-    Json.fromJson[AiBotSuggestReplyRequest](json)
-  }
-
-  def serializeResponse(response: AiBotSuggestReplyResponse): JsObject = {
-    Json.toJsObject(response)
-  }
-
+object AiBotSuggestReplyResponse {
+  def from(accountId: AccountId, results: String): AiBotSuggestReplyResponse =
+    AiBotSuggestReplyResponse(
+      accountId = accountId,
+      suggestion = results)
 }
+
+case class AiBotSuggestReplyResponse(
+                                      accountId: AccountId,
+                                      suggestion: String
+                                    )extends WithAccountId
