--- tmail-backend/tmail-third-party/ai-bot/src/test/resources/usersrepository.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one   
-  or more contributor license agreements.  See the NOTICE file 
-  distributed with this work for additional information        
-  regarding copyright ownership.  The ASF licenses this file   
-  to you under the Apache License, Version 2.0 (the            
-  "License"); you may not use this file except in compliance   
-  with the License.  You may obtain a copy of the License at   
-                                                               
-    http://www.apache.org/licenses/LICENSE-2.0                 
-                                                               
-  Unless required by applicable law or agreed to in writing,   
-  software distributed under the License is distributed on an  
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       
-  KIND, either express or implied.  See the License for the    
-  specific language governing permissions and limitations      
-  under the License.                                           
- -->
-
-<!-- Read https://james.apache.org/server/config-users.html for further details -->
-
-<usersrepository name="LocalUsers">
-    <algorithm>PBKDF2</algorithm>
-    <enableVirtualHosting>true</enableVirtualHosting>    
-    <enableForwarding>true</enableForwarding>
-</usersrepository>
-
